added a main function which takes arguments for initialization/configuration of user data added corresponding methods made templates for the user data changed way apis are saved in json (todo: change this when reading them!)

+++ install_and_init_bot_here.bat
@@ -1,4 +1,4 @@
-python -m pip install eazebot
 cd %~dp0
+if not exist "eazebot/EazeBot.py" python -m pip install eazebot
 python -c "from eazebot.auxiliaries import copy_user_files; copy_user_files()"
 pause
